rust: update dependencies, bumping fastcdc, russh and the vhost stack
Refresh Cargo.lock to the latest semver-compatible versions of every
dependency, which also clears the yanked spin 0.9.8 warning.

Move the crates behind a new major to it and adjust the call sites their
new APIs require: fastcdc 3->4, russh 0.61->0.62, and the vhost-user
virtio-fs stack (vhost, vhost-user-backend, virtio-queue, vm-memory,
vmm-sys-util). The bundled virtiofsd keeps its vm-memory distinct from
libkrun's pinned copy, re-wrapping slices across the two by pointer.

+++ vk-driver/src/registry.rs
@@ -37,9 +37,9 @@ use crate::image::{self, Reference, ResolvedImage};
 use crate::jobctx::JobCtx;
 
 // CDC parameters for runner.ext4 (FastCDC v2020): min 1 MiB, avg 4 MiB, max 16 MiB.
-const CDC_MIN: u32 = 1 << 20;
-const CDC_AVG: u32 = 4 << 20;
-const CDC_MAX: u32 = 16 << 20;
+const CDC_MIN: usize = 1 << 20;
+const CDC_AVG: usize = 4 << 20;
+const CDC_MAX: usize = 16 << 20;
 // Fixed zstd level: identical raw chunks must compress to identical bytes for the
 // blob digest (sha256 of the compressed bytes) to dedup. Shared with regserve's
 // transparent storage compression.
